Moved fancysets to sets module
