+++ sympy/sets/__init__.py
@@ -1,0 +1,1 @@
+from fancysets import TransformationSet
